defect-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
tensorflow==2.14.0
numpy==1.24.3
opencv-python==4.8.0.74
Pillow==10.0.0
Werkzeug==3.0.0
